+++ Parser/TokenFactoryInterface.php
@@ -29,6 +29,5 @@ interface TokenFactoryInterface
     public function setRule($rule);
     public function setClass($class);
     public function setBase($base);
-    public function setOffset($offset);
     public function setType($type);
 }